Move 'onClickSave' processing logic from EmailRecipients to ManageEmailGroups
